- Moved GrabScan with the other scan classes
